move tests to plug
